--- src/forms/index.ts
@@ -1,19 +1,0 @@
-/**
- * @packageDocumentation
- * @module Module @wesib/generic/forms
- */
-export * from './adjacent-field';
-export * from './field';
-export * from './field.share';
-export * from './field-name.amendment';
-export * from './form';
-export * from './form.share';
-export * from './form-preset';
-export * from './form-scope';
-export * from './form-unit';
-export * from './presets';
-export * from './on-submit.amendment';
-export * from './scoped-form-config';
-export * from './shared-field.amendment';
-export * from './shared-form.amendment';
-export * from './shared-form-unit.amendment';
