--- src/forms/presets/index.ts
@@ -1,3 +1,0 @@
-export * from './abstract-form-preset';
-export * from './form-css.preset';
-export * from './form-mode.preset';
